CAN/requirements.txt
tqdm
getch
numpy
serial
serial
folium
pyserial
termcolor
python-can
opencv-contrib-python